Test the BidomainSolver for replay, and tlm/adjoint Taylor convergence.
Getting this test working was quite difficult, because of parameters polluting the next tests. It seems that PETSc remembers some solver parameters deep inside, which are then used in the next test (resetting the dolfin.parameters does not help here). So you have to be extremly careful with the solver parameters you are using.